created components directory and moved app to there
